feat: write a multiplier, pass tests when signed
Currently, it has only one cycle delay, hence the
combinational delay may be too large.

It has eight 3-2 CSAs and four 4-2 CSAs, which can
handle signed-signed, signed-unsigned, and
unsigned-unsigned multiplication in theory, but yet
to be tested for the later two circumstances.

More work should be done to divide it into more
than a cycle and to go through more test cases.