feat: implement `A` extension instrs except lrsc
As we currently only use a hart, we need not consider
about inter-processor atomic issues, but it is not
difficult to implement. To be more specific, supposing
that we do not introduce disorder to memory accessing,
we may simply ignore the `aq` and `rs` bit in `amo`
instructions, and may use optimistic locking mechanism
to ensure the atomicity, although starvation problem
may exist.

But then again, we need not implement that stuff right
now.